recomendador/requirements.txt
nltk==3.8.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.15.0
pyodbc==4.0.39
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sklearn==0.0.post1
stack-data==0.6.2